requirements.txt
Django>=4.2.7,<5.0
djangorestframework>=3.14.0,<4.0
django-cors-headers>=4.3.1,<5.0
djangorestframework-simplejwt>=5.3.0,<6.0
python-dotenv>=1.0.0,<2.0
psycopg2-binary>=2.9.0,<3.0
Pillow>=10.0.0,<11.0
